[Refactor] Context API 를 이용해 관심사 분리
알림에 관한 상태를 NotificationProvider 안으로 이동시켜 props 대신 Context API를 이용하도록 수정
알림에 관한 상태를 구독하고 있던 컴포넌트들은 props 에서 Context API로만 바뀐것이기 때문에 Context API로 인한 불필요한 리렌더링은 발생하지 않음
